:books: Clean up the installation section of the README
Make it clear which installation methods are provided by me, and which are
provided by third parties (I welcome such contributions, of course, but
can't vouch for them so I think it's prudent to make this clear).

Also remove `$` from all of the `sh` code blocks (including my own) so it's
easier for people to copy/paste commands.