:sparkles: Add some vim/less-a-like binds to the document viewer
I'm not a vim user, never have been; I can quickly load, edit and save stuff
with it in an emergency, but normally I'm an Emacs user like the gods
intended. I use less fairly often but seldom live in it long enough to
really have the muscle memory (if I need to really read something I'll load
it in Emacs of course!).

So this is my attempt to make the document viewer widget itself a little
more friendly to all.