:zap: Switch from rye to uv for package management
Another run at #123 with thanks to @hugovk for the initial PR and the push
to get to grips with this.

This commit is pretty much the same as was done in that PR, and carries over
a lot of code. However I initially had to back it out because it killed my
ability to publish Hike.

Now I've done a much of reading I think I'm confident that it can be used
without too much effort. That is, I think I've managed to gain the minimum
amount of trust I need in it to be able to make the switch.

It's... not as good as rye, as I'm seeing it so far, but advantages exist
and I can work around the friction that uv introduces.

Of course the main motivation is that the bait and switch of rye->uv has
finally happened. It's yet another reason for folk to view Python packaging
tools with deep (and justified) suspicion.